Dodawanie+usuwanie kategorii

+++ app/src/main/java/com/example/quicknote/data/Note.kt
@@ -10,7 +10,7 @@ import java.io.Serializable
  * @property id Unikalny identyfikator notatki (klucz główny, autogenerowany).
  * @property title Tytuł notatki.
  * @property content Treść notatki.
- * @property categoryId Identyfikator kategorii (0=Brak, 1=Praca, 2=Szkoła, 3=Dom, 4=Inne).
+ * @property categoryId Identyfikator kategorii powiązany z Category.id.
  * @property priority Poziom priorytetu (0=Zwykła, 1=Ważna).
  * @property createdAt Znacznik czasu utworzenia notatki (w milisekundach).
  * @property reminderTime Czas zaplanowanego przypomnienia (w milisekundach, 0 jeśli brak).
@@ -21,7 +21,7 @@ data class Note(
     val id: Long = 0,
     val title: String,
     val content: String,
-    val categoryId: Int = 0,
+    val categoryId: Long = 0,
     val priority: Int = 0,
     val createdAt: Long = System.currentTimeMillis(),
     val reminderTime: Long = 0
